backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic[email]==2.5.2
